Cache: If LRUStorage is used, create ST real storage
If the underlying storage does not support max_count and/or max_size
and it accordinly is decorated with LRUStorage, then create the real
storage as single-threaded. Since LRUStorage will do locking it is of
no use to do locking in the real storage as well.

+++ server/modules/filter/cache/storagefactory.hh
@@ -64,7 +64,7 @@ public:
     Storage* createStorage(cache_thread_model_t model,
                            const char* zName,
                            uint32_t ttl,
-                           uint32_t max_count,
+                           uint64_t max_count,
                            uint64_t max_size,
                            int argc, char* argv[]);
 
@@ -91,7 +91,7 @@ public:
     Storage* createRawStorage(cache_thread_model_t model,
                               const char* zName,
                               uint32_t ttl,
-                              uint32_t max_count,
+                              uint64_t max_count,
                               uint64_t max_size,
                               int argc, char* argv[]);
